vioscsi: always kick VQ when buffers exhausted
When VIRTIO_RING_F_EVENT_IDX is not enabled, we should kick the VQ
even if VIRTQ_USED_F_NO_NOTIFY is set when all buffers are exhausted.

This will ensure that we always kick when we're full, not only *after*
we are full. Do this by keying off of vq->num_unused <=1 in
virtqueue_kick_prepare_split.

On the vioscsi side, after successfully adding a buffer to VQ, query
virtqueue_kick_prepare under lock to ensure that we can get an accurate
read of vq->num_unused.

nit: cleanup virtqueue_kick_prepare_split to only deal with
event_suppression_related variables if that is enabled.

Signed-off-by: Jon Kohler <jon@nutanix.com>