viosock: Use STATUS_IO_TIMEOUT instead of STATUS_TIMEOUT
The trouble with STATUS_TIMEOUT is that it is considered being a success code (the NT_SUCCESS macro evaluates to TRUE). This may be important since this macro is very often used to check for success in kernel code. On the other hand, STATUS_IO_TIMEOUT is traded as an error code by this macro, so let's use it to properly indicate timeout error condition.

Signed-off-by: Martin Drab <martin.drab@virtuozzo.com>