notify-obj: detach netkvmno.dll from the VIOPROT protocol
The notification is not needed now, the logic of the operation is
changed and implemented in the protocol service. To avoid changes
in packaging we do not change the set of files but the DLL does
not service the protocol anymore.
Future plan is to remove also protocol binary from the protocol
installation and move it to the installation of netkvm.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>