WSK: Implement WskListen
This call is considered as advanced one since there is no need to call "listen" in the WSK wolrd. To create a listening socket, one just needs to specify the WSK_FLAG_LISTEN_SOCKET in a call to WskSocket. Thus, WskListen can be used just to change the connection backlog parameter.

Signed-off-by: Martin Drab <martin.drab@virtuozzo.com>