vioscsi: add DeviceReset on CompletePendingRequests path
We discovered that if a backend storage eventually completes a write IO
request that was completed by vioscsi on the LUN reset handler, it may
lead to inconsistency between data written in the backend and the expected
data on the Windows side.

In case if it such an IO request was in-paging write request, Windows MM
will detect it as CRC error and will bugcheck the guest with 0x1A bugcheck.

Adding DeviceReset() should try to complete IO requests in the backend,
reducing the chance to hit the abovementioned inconsistency.

Signed-off-by: Sergey Bykov <sergey.bykov@nutanix.com>