[vioser]: Do not wait infinitely when power down or removal is imminent
This commit prevents the Virtio Serial Driver from potentially infinitely waiting for the hypervisor when switching to the "WdfPowerDeviceD3Final" state (that happens when the system is being shut down, or the device is being removed). In such case, the XxxEvtDeviceD0Exit callback routine must finish within certain time interval. That is achieved by enhancing VIOSerialSendCtrlMsg by an extra parameter determining whether long waits are possible.

Signed-off-By: Martin Drab <martin.drab@virtuozzo.com>