[viogpu]: Cope with larger than expected DXGKRNL_INTERFACE size on StartDevice
This PR deals with the situation when the DXGKRNL_INTERFACE passed to VioGpuDod::StartDevice is larger than driver's place for such structure, represented by the m_DxgkInterface attribute. In such a case, the driver just degrades the structure to the largest it can handle. This should be fine since DXGKRNL_INTERFACE only adds new members when advancing to its new versions.

Fixes: #967
Signed-off-By: Martin Drab <martin.drab@virtuozzo.com>