viogpudo: fix disappearing cursor issue
`SetPointerShape` calls `CreateCursor` which copies the shape data to
the host via `TransferToHost2D`. Then it calls `QueueCursor` which sets
the cursor on the host. There is a race condition between these two
functions because `TransferToHost2D` uses the `m_CtrlQueue` and
`QueueCursor` uses the `m_CursorQueue`. Whenever the cursor queue is
handled first by QEMU, the host attempts to set the cursor image by
reading the buffer that is still empty because `TransferToHost2D` had
not yet executed.

The fix here is to wait on `TransferToHost2D` to complete before calling
`QueueCursor`. We achieve this by waiting on an KEVENT and moving some
code around in the IRQ handler to set this event after recieving the
response for the transfer command.

Signed-off-by: Joelle van Dyne <joelle@turing.llc>