vioserial : host connection status set to false when driver exit
In some scenario such as suspend the guest, the host will notify connect close,
but the status can not be sync to guest driver,
which will cause a bug while the guest write the seiral before the host connect happend.
so it's better just let the connection status to be false rather then true.
because the next boot(resume) will enable the connection again

Signed-off-by: zeroway <M_O_Bz@163.com>