netkvm: remove query for OID_GEN_TRANSMIT_QUEUE_LENGTH
https://issues.redhat.com/browse/RHEL-19441
The response to the query is calculated incorrectly.
We can make it correct but I think that support for this
OID is not required, it is not in the list of mandatory
OIDs. In general the NDIS has the same knowledge about
how many packets are queued by the miniport.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>