netkvm: change limit of RX buffers per queue 256=>1024
https://issues.redhat.com/browse/RHEL-19430

The number of buffers is always limited by queue size,
so with the default QEMU setting (256) the driver will do
the same. If QEMU is configured in command-line to have
bigger queue (it allows up to 1024), the previous driver
required to change also driver property to really use
full size of QEMU RX queue. Now this will not be required
(up to 1024 buffers).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>