netkvm: request to respawn the DPC if TX queue is not empty
In the crash of Mini6RSSSendRecv test found a situation
when there is one packet is still in the SendQueue when
the protocol driver waits for unfinished packet completion.
Try to solve it by requesting additional DPC also in case
when the TX queue is not empty (currently we request it
only when there are completed packets in Tx queue and we
can't enable interrupts for the queue).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>