[viosock]: Take care of recv requests stucks after socket shutdown
When an other end (peer) of the socket reports a shutdown, the recv request queue is not processed. This means that some recv calls may get stuck forever. This commit changes the behavior -- when a shutdown from its peer is received, the recv queue gets processed and all requests that would get stuck are completed with STATUS_SUCCESS (which usually ends in receiving zero bytes).

Signed-off-by: Martin Drab <martin.drab@virtuozzo.com>