netkvm: add getter for current OS NDIS version
This additional getter allows to query current NDIS version
of the OS regardless whether we compile the driver for such version
or not. For example: the driver is built for NDIS 6.85, it runs on
Server 2025 (NDIS 6.89 or higher), effective version will be 6.85
but in the code we are able to check whether the OS supports 6.89.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>