[viosock]: Introduce the WSP service
This service is responsible for keeping the WSP for the Viosock driver installed as long as it is running (i.e. it installs the WSP during its initialization and removes it on its stop action). This approach seems to be the best/simplest regarding WSP installation from the INF file on Windows 10 since:
- co-installer is no longer supported by newer WDKs,
- the AddSoftware INF directive is not supported by some old (but before EoL) versions of Windows 10,
- the co-installer is prohibited regardless of the Windows version (i.e. the WDK complains even if corresponding INF section is used only on old versions of Windows).

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>