[CI-NO-BUILD] Add interruption handler to CollectSystemInfo
Add an interruption handler. When the script is not finished
running, the generated file name could be marked with the
ongoing process, such as Collecting_Status.txt. and it would
be delete after completion. In this way, when the script is
waiting to be interrupted, it will be clearer to the user
which directory was interrupted.

Solves: https://issues.redhat.com/browse/RHEL-50169
Signed-off-by: Dehan Meng <demeng@redhat.com>