[viosock]: Do not unregister event on WSAEnumNetworkEvents call
The documentation states: The recording of network events commences when WSAEventSelect is called with a nonzero lNetworkEvents parameter and remains in effect until another call is made to WSAEventSelect with the lNetworkEvents parameter set to zero, or until a call is made to WSAAsyncSelect..[1]

Thus, let's not disassociate the event object when WSAEnumNetworkEvents is called. Just reset it into non-signaled state.

Also, this commit fixes reference count leak from translating the event handle to KEVENT pointer.

ref: https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-wsaenumnetworkevents

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>