virtio-fs: retrieve maximum queue size
Currently the driver does not set any limit for the number
of scatter-gather entries for virtio messages regardless
of actual queue size (default size is 128, max is 1024).
In case of non-indirect transfer (it is always used for Q0,
'forget' requests) this does not cause spec violation
as overfragmented messages will fail on add_buf.
In case of indirect transfers (Q1, most of requests) the driver
places up to 512 fragments which might be a spec violation.
Currently we do not set the limit in legacy (non-WDF) code,
just keep queue size for later usage.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>