virtio-fs: increase indirect area from 8K to 16K
This is single allocation used for Q1. Typical (recommended
although non-default) setting of virtio-fs device includes
queue size of 1024, so we align the indirect area according
to this number. 1024 is also the maximum number of SG
elements that virtiofsd supports and the maximum size of the
virtqueue that QEMU allows for this device.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>