virtio-fs: extend printouts related to request cancellation
The code related to request cancellation follows KMDF documentation.
I suspect the documentation might be partially incorrect and
there is a possibility for race condition:
if just before WdfRequestUnmarkCancelable the cancel happens
and completes the request the WdfRequestUnmarkCancelable will
not return STATUS_CANCELLED (in the best case it will return
STATUS_INVALID_DEVICE_REQUEST) and the driver will still try
to complete this request.
Added TODO and more information to log.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>