viofs: accept VIRTIO_F_ACCESS_PLATFORM in future
https://issues.redhat.com/browse/RHEL-11405
Currently the device never suggest this feature due to backend
limitation. If we clear it intentionally in the driver, the driver
will not be able to start in future when the device will support
IOMMU.
In previous versions it was correct to clear this flag to avoid
DMA translation as the driver did not use translation-compatible
DMA API. Now, after switch to WDF DMA API we can accept this
feature in the driver, making it forward-compatible.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>