viogpusc: Fix the timeout issue when stopping viogpusc service
When the user stops the viogpusc service either through SCM or
console command line, a timeout error is thrown out as the
following,

From the console,
The service did not respond to the start or control request
in a timely fashion.

From SCM
A timeout was reached while waiting for a transaction response
from the viogpusc service.

This error is due to the incomplete initialization of the
Session creating viogpuap process. This initialization is
blocked until the process is terminated. When SCM or console
command tries to stop the service, it always fails because the
Session is waiting for the process to be terminated. Actually,
the process is waiting for the service to trigger an event to
terminate itself. This is a dead loop.

To fix this issue, the Session initialization should be completed,
and the Session Manager should monitor the termination of the
process and release the resources after the process is terminated.

Signed-off-by: Annie Li <annie.li@oracle.com>