netkvm: deregister miniport in erroneous path
https://issues.redhat.com/browse/RHEL-68725

The problem can be reproduced under driver verifier with
low resource simulation. If the miniport is registered
but the protocol fail to register we need to deregister
the miniport. Otherwise in future attempts to load the driver
the NDIS enters erroneous state and may issue BSOD. Example of
BSOD: BUGCODE_NDIS_DRIVER/NDIS_BUGCHECK_REFCOUNT_IMBALANCE.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>