Fix IntLock regression in commit 15e64ace.
This corrects interrupt/spinlock acquision code regressed
previously. Locking in the driver is now the same as
prior to that point, and basic functionality is restored.

Initial regression testing has been performed. There are
however, some outstanding questions as to whether the driver
performs correctly in the crashdump/hibernation path.

I will be performing additional regression testing for
those cases, and MSI(X), and issuing further patches
if necessary.

Ref: Nutanix ENG-741981.

Signed-off-by: Martin Harvey <martin.harvey@nutanix.com>