viogpu: Allocate more buffers than the queue capacity
I occasionally saw viogpu crashes the whole system because of AllocCmd()
returning NULL. viogpu allocates buffers according to the queue size and
while it works fine in most cases, the specification says:
> Once the ring is full of driver descriptors, the driver stops sending
> new requests and waits for the device to start processing descriptors
> and to write out some used descriptors before making new driver
> descriptors available.

While waiting for the device in such a scenario, the driver will need to
hold buffers more than the queue size.

Allocate a buffer if we need more than the queue size, but free it once
it becomes unnecessary so that we will not keep unbound amount of memory
indefinitely.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>