[viostor]: Use the Status table to hold the command header as well
When a pending request gets completed by the driver before it is handled by the host, it may happen that the host have not read the command header yet. This means a read from a physical memory no longer owned by the driver. Moving the physical memory holding the command header to the Status table as well enforces that the physical memory is always owned by the driver when the host code is reading from it.

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>