netkvm: maintain NBL chain information in CNBL
The new logic to support chains is placed under ifdef currently
and the driver works in backward compatible mode when all the
procedures related to chain information in the NBL are stubs.

The new logic that takes in account NBL chain is:
For single NBL (not chained) - all 3 fields in m_Chain are 0.
For first NBL in chain:
    m_FirstInChain = 0
    m_NumChained is number of NBLs following the first one
    m_NumCompleted is incremented on each NBL completion
For following NBLs in chain:
    m_FirstInChain is CNBL with first NBL
    m_NumChained is serial number in chain
    m_NumCompleted is not used (0)
Each chained CNBL before mapping AddRef's the first one on
    creation and Release's it in destructor, so lifetime of
    the first CNBL is alive till one or more of chained ones
    alive.
The condition when the CNBL is done and can be released and
    completed:
    Single NBL - as before (all NBs are completed)
    First in chain - the same and all the chained NBLs are done
    Chained - the same and the first one is done
Such a way, all the CNBLs for chained NBLs become 'done' at the
    same moment

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>