netkvm: recreate the NBL chain on CNBL completion
Now we enable NBL chain support.
Deliver entire NBL chain if one was provided in Send() call.
Each CNBL created for each 'next' NBL (following to the first
one) on completion of all it's NBs detaches it's NBL and
attaches it to the m_NBL of the first one. Such a way when
all the CNBLs related to the same chain are ready, only first
one has m_NBL (with all the rest of NBLs chained to it).
For all the rest of CNBLs the m_NBL is NULL.
The order of NBLs in the chain might be slightly different
than original one (due to NB completions running simultaneously
in different flows), for that the chain is tested and repaired
if needed before it is indicated up.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>