netkvm: extend structure RxNetDescriptor
Add new fields for flexibility of:
- page number where we have a virtio header (currently 0)
  offset of data start in page 1 (currently 0)
- index of page number (currently there is no difference
  between page index and index in SG array)
Currently these fields are not in use. In future, in case
the header and data are combined:
the header page will be 1
the data start offset will be size of vnet header
the first page will not be used by SG, it will contain
only indirect area and the tail of the data.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>