[RHEL-71813] netkvm: enable combining vnet header and data
Jira link: https://issues.redhat.com/browse/RHEL-71813

If the device is V1 or suggests VIRTIO_F_ANY_LAYOUT it does
not have any assumptions regarding framing of packets and
does not require to have separate descriptor for vnet
header. On bare metal device this reduces number of DMA
transaction in RX path.
In order to get back to previous behavior it is enough to
revert this commit or just set combineHeaderAndData to
be false always.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>