netkvm: do not adjust the MDL length
First, this is not needed. The exact size of the payload
is defined in the NB and the MDL chain may be longer.
Second, this complicates the configuration of the first
MDL when we need to combine vnet header and data.
If we change the length of MDL we need to restore it upon
deletion, so we need to save the initial length (it might
be different from the size of respective page if the data
and header are combined).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>