netkvm: move getting control response to separate procedure
For further reuse we move the code to separate procedure
without changing the behavior.
To be able to retrieve the response status we also save
the offset where the response should be before submitting
the command. This does not change anything as we set this
offset under m_Lock and use it under the same lock.
Several TODOs will be addressed in future commits.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>