netkvm: split SendControlMessage procedure
Add internal procedure SendInternal, it is called under lock
and will be used in future to resend queued messages that
came when the control queue was busy.
In general, we need to know whether this is an initial call
for the command or processing of queued command.
In the first case we should schedule the command if the queue
is busy or if there are already queued command.
In the second case we shouldn't schedule the command.
Timeout for getting the response in the first case might be
longer.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>