[RHEL-88155] netkvm: allow control command timeout
Maintain status of pending command and list of commands for
delayed execution. If the response is not received during
reasonable time, assume the device still functions and will
respond later. Recheck when the next command is requested
and when CX interrupt comes (probably as the result of the
response).
To avoid misunderstandings only GetResponse can modify the
m_PendingCommands, all other can only check the status and
call GetResponse when needed.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>