[RHEL-71813] netkvm: add option to try allocating contiguous RX memory
Jira link: https://issues.redhat.com/browse/RHEL-71813
Fixes: #1078

If "Init.SeparateRxTail" is set to 0 the driver will try to
allocate contiguous memory block for each RX buffer. If it
succeeds, the virtio descriptor will contain single entry
and the device will not spend time retrieving the indirect
chain. For the bare metail device this might be better.
We leave the default value "1" and by default the driver
will pace the tail in separate memory block (where we
keep indirect descriptors).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>