netkvm: add accesor for max usable data in TX headers area
Each TX descriptor includes 4K block when we place the
virtio header, ethernet header, VLAN/Priority tag and
(when needed) the IP and TCP header that the driver
can modify before submitting to the TX queue. The rest of
the block the driver can use for packet data copy if
needed. Keep the size of this free area and add getters
to TX virtqueue and TX path objects.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>