netkvm: add a flag to be able to do TX without mapping
Nothing is changed yet. We just add the flag in the NBL
that will allow us in future to skip the mapping if all
the packet data can be copied to the headers block.
We set this flag when the packet size allos that and when
the packet is not LSO,CSO,USO,IP header checksum is not
required and 'any layout' is accessible by the device.
Typically this flag will be set for non-IP packets.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>