netkvm: work without packet mapping if allowed
https://issues.redhat.com/browse/RHEL-78518
https://issues.redhat.com/browse/RHEL-82071

If the flag m_SkipMapping is set in the NBL, do not
request creation of NBL for it. Copy entire packet into
headers area in the descriptor and transmit the packet
using single TXQ descriptor or two (if VLAN/Prio tagging
is required). This will be applied mainly to non-IP packets.
Such effective solves the problem of 1c_Mini6PerfSend test
on Win10 when the DMAR is enabled. This minimizes the amount
of work the driver verifier should do per packet.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>