[vioser]: Do not requeue write requests completion of which is imminent
When a device stop is imminent (temporary one, not due to device removal), the Virtio Serial driver uses WdfRequestStopAcknowledge to requeue pending write requests to its write queue. This operation changes ownership of such requests from the driver to WDF. However, the driver may still attempt to complete the requests during its DPC processing. This is not allowed since a driver can only complete requests it owns. WDF Verifier complains about that.

This commit changes the behavior the following way:
- Write requests still present in the WriteBuffersList list are detached from the list (i.e. the Request pointer in the corresponding list entry is set to NULL) and requeued. The detach operation prevents their completion when owned by WDF.
- Write requests not present in the WriteBuffersList list are ignored (not ackowledged) since their completion is imminent. DPC is in progress and already removed them from the list and is about to complete them. Ignoring such requests in the EvtIoStop callback makes WDF to wait for their completion before continuing with device's power-down sequence.

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>