[vioser]: Handle QueryRemove PnP event to better cooperate with QGA
When QGA detects a Virtio serial device exists, it automatically opens a handle to it and submits a read request to receive future data from the host. This connection basically exists as long as the QGA executable is running. Unfortunately, the open handle and pending read request prevents the driver and its devices from removal (device disabling, driver update) – the device remains stuck in its removal phase until the handle gets closed (meaning forever).

This commit enhances the driver with a QueryRemove event handler. The handler is invoked when the PnP Manager asks whether the device can be removed. The handler completes the pending read request with STATUS_END_OF_FILE which alerts the QGA service. Currently, QGA just resubmits the read request again, however, its code will be modified to close all handles to Virtio serial devices for a while in order to allow the devices to get removed. That means, adding the QueryRemove handler does not change functionality of QGA.

This commit does not affect temporary device stop due to hardware resource reassignment since that is already working fine.

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>