[vioscsi] honor seg_max from the device when calculating max_physical_breaks #1456
The seg_max field in the config space of a virtio-scsi device indicates
the maximum number of segments (aka physical breaks, aka I/O vectors)
it can handle in a single command, not accounting for the two I/O
headers.

The vioscsi driver must never send more segments in a single command
than the device announced it can handle.